% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_Beta.R
\docType{data}
\name{Example_Beta}
\alias{Example_Beta}
\title{Example Beta to test for HiTIMED}
\format{
The matrix contains methylation beta values

        The format is:
        num [1:485512, 1:3] 0.90508063  0.52818155  0.63286658 ...
}
\usage{
Example_Beta
}
\description{
This object contains a matrix of beta values from three colon 
    adenocarcinoma samples 6285625076_R06C02", "6929718079_R01C01", 
    "6042316015_R02C02". These samples are from publicly available 
    data source TCGA at https://portal.gdc.cancer.gov/.
}
\examples{
data("Example_Beta")
head(Example_Beta)
}
\keyword{datasets}
