% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExTIME_deconvolution.R
\name{ExTIME_deconvolution}
\alias{ExTIME_deconvolution}
\title{ExTIME_deconvolution}
\arguments{
\item{tumor_beta}{Methylation beta matrix from the bulk tumor samples.}

\item{tumor_type}{Specify tumor type for microenvironment deconvolution.}

\item{h}{Specify the layer of deconvolution in the hierarchical model. Default is 6.}
}
\value{
A matrix with predicted cell proportions in tumor microenvironment.
}
\description{
The function estimates proportions up to 17 cell types in tumor
microenvironment for 20 types of carcinomas.
}
