% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiTIMED_deconvolution.R
\name{HiTIMED_deconvolution}
\alias{HiTIMED_deconvolution}
\title{HiTIMED_deconvolution}
\usage{
HiTIMED_deconvolution(tumor_beta, tumor_type, h = 6, tissue_type = "tumor")
}
\arguments{
\item{tumor_beta}{Methylation beta matrix or data frame from the bulk tumor samples.}

\item{tumor_type}{Specify tumor type for microenvironment deconvolution.
BLCA: Bladder urothelial carcinoma, BRCA: Breast invasive carcinoma,
CESC: Cervical squamous cell carcinoma and endocervical adenocarcinoma,
CHOL: Cholangiocarcinoma, COAD: Colon adenocarcinoma, ESCA:
Esophageal carcinoma, HNSC: Head and neck squamous cell carcinoma,
KIRC: Kidney clear cell renal cell carcinoma, LIHC:
Liver hepatocellular carcinoma, LUAD: Lung adenocarcinoma,
LUSC: Lung squamous cell carcinoma, OV: Ovarian addenocarcinoma,
PAAD: Pancreatic adenocarcinoma,
PRAD: Prostate adenocarcinoma, READ: Rectum adenocarcinoma,
STAD: Stomach adenocarcinoma, THCA: Thyroid carcinoma,
UCEC: Uterine corpus endometrial carcinoma}

\item{h}{Numeric variable. 
Specify the layer of deconvolution in the hierarchical model. Default is 6.}

\item{tissue_type}{specify whether the tissue is tumor. Default is tumor. If not
tumor, the function will preset the tumor purity to 0.}
}
\value{
A matrix with predicted cell proportions in tumor microenvironment.
}
\description{
The function estimates proportions up to 17 cell types in tumor
microenvironment for 20 types of carcinomas.
}
\examples{
#Step 1: Load example data
library(ExperimentHub)
Example_Beta<-query(ExperimentHub(), "HiTIMED")[["EH8092"]]
#Step 2: Run HiTIMED and show results
HiTIMED_result<-HiTIMED_deconvolution(Example_Beta,"COAD",6,"tumor")
head(HiTIMED_result)

}
